requirements.txt
streamlit
transformers
faiss-cpu
sentence-transformers
python-docx
python-pptx
pandas
PyMuPDF
accelerate
torch
